requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pip==23.3
pluggy==1.3.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
selenium==4.14.0
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
wheel==0.41.2
wsproto==1.2.0
